Matching with time-varying treatment

Matching with time-varying treatment

Usage

tvt_date2numeric(
  data,
  date_statrisk,
  date_treatna,
  date_eos,
  state_eos,
  date_eoatrisk = NULL
)

tvt_matching_date(
  data,
  date_statrisk,
  date_treatna,
  date_eos,
  state_eos,
  date_eoatrisk = NULL,
  ratio = 1,
  by = NULL
)

tvt_matching_num(
  data,
  time_atrisk,
  time_eos,
  state_treat,
  state_eos,
  ratio = 1,
  by = NULL
)

tvt_summary_cases(
  data,
  date_statrisk,
  date_treatna,
  date_eos,
  state_eos,
  date_eoatrisk = NULL,
  by = NULL
)

Arguments

data

a data frame

date_statrisk

column names with the date when patients start to be at risk of the treatment.

date_treatna

column names with the date of the start of the treatment and NA's for untreated subjects.

date_eos

column names with the date of the end of the study

state_eos

column names indicating the state at the end of the study

date_eoatrisk

column names with the date when patient end to be "at risk of the treatment". Specify date_eos the ending of the "at risk of the treatment" period corresponds to the end of the study.

ratio

Integer. The controls per cases.

by

column names indicating a grouping factor for the matching.

time_atrisk

column names with integer representing the length of the "at risk of the treatment" period.

time_eos

column names with integer representing the length of the study.

state_treat

column names indicating if the subject experimented the treatment

Value

The tvt_summary_cases() function return a tibble with information of control candidates for each case. The tvt_date2numeric() convert the encoding of the event from dates to numeric. The tvt_matching_date() and tvt_matching_num() perform the matching algorithm.

Examples

data("tvtdata")
tvtdata_num <- tvt_date2numeric(tvtdata,stos,treatna,eos,eos01)
tvt_sum <- tvt_summary_cases(tvtdata,stos,treatna,eos,eos01, by = fct)

set.seed(42)
tvt_match<- tvt_matching_date(tvtdata,stos,treatna,eos,eos01)

set.seed(42)
tvt_match2 <- tvt_matching_num(tvtdata_num,time_atrisk,time_eos,state_eos,state_treat)